Rename DetailMinimap to shared Minimap
    Move util to utils/minimap.ts so it is not gallery-scoped.
