+++ utils/minimap.ts
